Revert 628e26d..f57c299
This rolls back to commit 628e26d7d08c5c2cad2249760849135c371f4f12.

+++ src/Game.h
@@ -63,7 +63,7 @@ public:
      *
      * @num is the value to update by.
      */
-    void updateHealth(double value);
+    void updateHealth(float value);
 
     /**
      * Updates the gold counter.
